+++ CHANGELOG.md
@@ -29,7 +29,8 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ### Fixed
 - Bump MySQL to 5.6.42. Resolve [JPERF-345].
 - Set up VUs only once per node in `MulticastVirtualUsers`. Fix [JPERF-346].
- 
+- Chromedriver continues to work after a page load timeout. Fix [JPERF-349].
+
 ### Added
 - Separate Chromium implementations, like `Chromium69` or `Chromium70`. Resolve [JPERF-350].
 - Install Chromium with a compatible chromedriver. Resolve [JPERF-352].
@@ -44,6 +45,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 [JPERF-346]:https://ecosystem.atlassian.net/browse/JPERF-346
 [JPERF-350]:https://ecosystem.atlassian.net/browse/JPERF-350
 [JPERF-352]:https://ecosystem.atlassian.net/browse/JPERF-352
+[JPERF-349]:https://ecosystem.atlassian.net/browse/JPERF-349
 
 ## [4.5.1] - 2019-01-04
 [4.5.1]: https://bitbucket.org/atlassian/infrastructure/branches/compare/release-4.5.1%0Drelease-4.5.0